Updates for Visual Studio

+++ src/test/ByteMap.test.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
-#include <gtirb/Constants.hpp>
 #include <gtirb/ByteMap.hpp>
+#include <gtirb/Constants.hpp>
 #include <gtirb/Utilities.hpp>
 #include <memory>
 
@@ -59,9 +59,9 @@ TEST_F(Unit_ByteMapF, legacy_byte)
     {
         const auto expectedWord = (((InitialByte & i) | ((InitialByte & (i + 1)) << 8)));
 
-        EXPECT_EQ(this->InitialByte & i,
-                  this->byteMap.getData8(Unit_ByteMapF::Offset
-                                         + gtirb::EA{static_cast<uint64_t>(i)}))
+        EXPECT_EQ(
+            this->InitialByte & i,
+            this->byteMap.getData8(Unit_ByteMapF::Offset + gtirb::EA{static_cast<uint64_t>(i)}))
             << "Bad byte read at : " << Unit_ByteMapF::Offset + gtirb::EA{i};
 
         if(i < this->InitializedSize - 1)
@@ -70,8 +70,8 @@ TEST_F(Unit_ByteMapF, legacy_byte)
 
             const auto word = this->byteMap.getData16(Unit_ByteMapF::Offset + gtirb::EA{i});
 
-            EXPECT_EQ(expectedWord, word) << "Bad word read at : "
-                                          << Unit_ByteMapF::Offset + gtirb::EA{i};
+            EXPECT_EQ(expectedWord, word)
+                << "Bad word read at : " << Unit_ByteMapF::Offset + gtirb::EA{i};
         }
     }
 }
@@ -79,8 +79,8 @@ TEST_F(Unit_ByteMapF, legacy_byte)
 TEST_F(Unit_ByteMapF, legacy_word)
 {
     this->byteMap.setData(gtirb::EA(0x401000), uint16_t{0xDEAD});
-    EXPECT_EQ(0xDEAD, this->byteMap.getData16(gtirb::EA(0x401000))) << "Bad word read at : "
-                                                                    << 0x401000;
+    EXPECT_EQ(0xDEAD, this->byteMap.getData16(gtirb::EA(0x401000)))
+        << "Bad word read at : " << 0x401000;
 }
 
 TEST_F(Unit_ByteMapF, legacy_dword)
@@ -111,8 +111,8 @@ TEST_F(Unit_ByteMapF, legacy_boundariesWithZero_0)
         }
         else
         {
-            EXPECT_EQ((this->InitialByte & (i - 16)), buf[i]) << "Bad chunk read at : " << ea
-                                                              << " plus : " << i;
+            EXPECT_EQ((this->InitialByte & (i - 16)), buf[i])
+                << "Bad chunk read at : " << ea << " plus : " << i;
         }
     }
 }
@@ -184,12 +184,11 @@ TEST_F(Unit_ByteMapF, legacy_sentinelSearch_1)
     }
 }
 
-
 TEST_F(Unit_ByteMapF, legacy_sentinelSearch_2)
 {
     // c. search for 254 -- like a, but two fewer
     const auto ea = Unit_ByteMapF::Offset + gtirb::EA{InitializedSize} - gtirb::EA{16};
-    const auto buf = this->byteMap.getDataUntil(ea, '\xfe', 16);
+    const auto buf = this->byteMap.getDataUntil(ea, static_cast<uint8_t>('\xfe'), 16);
     const auto bytesRead = buf.size();
 
     ASSERT_EQ(15, bytesRead);
